Implemented PR change  requests
