+++ Ginger/GingerCoreCommon/UIElement/IWindowExplorer.cs
@@ -1,4 +1,4 @@
-#region License
+﻿#region License
 /*
 Copyright © 2014-2023 European Support Limited
 
@@ -32,7 +32,7 @@ namespace Amdocs.Ginger.Common.UIElement
     {
         List<AppWindow> GetAppWindows();
         void SwitchWindow(string Title);
-        void HighLightElement(ElementInfo ElementInfo, bool locateElementByItLocators = false);
+        void HighLightElement(ElementInfo ElementInfo, bool locateElementByItLocators = false,IList<ElementInfo> MappedUIElements = null);
         void UnHighLightElements();
         string GetFocusedControl();
         ElementInfo GetControlFromMousePosition();
